+++ backend/lib/game-persistence/src/main/resources/schema.sql
@@ -99,8 +99,8 @@ CREATE TABLE competition_round_answers
     session_id   BIGINT NOT NULL,
     round_number INT    NOT NULL,
     team_id      BIGINT NOT NULL,
-    value        INT    NOT NULL,
-    income       INT    NULL,
+    value        FLOAT  NOT NULL,
+    income       FLOAT  NULL,
     PRIMARY KEY (session_id, round_number)
 );
